+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/NewEventWorkspace/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { NewEventWorkspace } from './NewEventWorkspace.component';